Support decimal chapter numbers
Convert chapter_number handling from integers to floats across the app and DB. Use parseFloat when reading chapter IDs and filtering, update ChapterReader navigation to find next/previous by index (instead of +/-1) so non-sequential decimal chapters work, and adjust ChapterForm input to accept decimal values (step="any", min=0) with parseFloat and regex updated to parse decimals. Add a migration to alter chapters.chapter_number to DOUBLE PRECISION to persist decimal chapter numbers.

+++ src/pages/NovelDetail.tsx
@@ -245,7 +245,7 @@ const NovelDetail = () => {
 
       if (isNumber) {
         // Exact match for chapter number
-        const num = parseInt(query);
+        const num = parseFloat(query);
         filtered = filtered.filter(ch => ch.chapter_number === num);
       } else {
         // Fuzzy match for title
